requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.4
python-dotenv==1.0.1
pydantic-settings==2.6.1
PyJWT==2.10.1
python-multipart==0.0.20